requirements.txt
#env
#uwsgi need C complier,such as gcc,g++,also python-devel
uwsgi==2.0.15
setuptools==40.8.0
pip==19.0.3

#app
django==1.11
MySQL-python==1.2.5
arundeckrun==0.2.3
celery==3.1.25
django-mail-templated==2.6.5
pycrypto==2.6.1
django-celery==3.2.2
paramiko==2.4.2

pika==1.0.1
celery-message-consumer==1.1.1
xlrd==1.2.0

python-dateutil==2.8.0

